Move renderComponentWithHook
