+++ apps/parsley/src/components/DetailsMenu/DetailsMenuCard/Toggles/PrettyPrintToggle/PrettyPrintToggle.test.tsx
@@ -3,10 +3,10 @@ import { MockInstance } from "vitest";
 import {
   act,
   renderWithRouterMatch as render,
+  renderComponentWithHook,
   screen,
   userEvent,
 } from "@evg-ui/lib/test_utils";
-import { renderComponentWithHook } from "@evg-ui/lib/test_utils/hooks";
 import { LogRenderingTypes } from "constants/enums";
 import { useLogContext } from "context/LogContext";
 import { logContextWrapper } from "context/LogContext/test_utils";
