+++ apps/spruce/src/context/toast/Toast.test.tsx
@@ -5,8 +5,8 @@ import {
   screen,
   userEvent,
   waitFor,
+  renderComponentWithHook,
 } from "@evg-ui/lib/test_utils";
-import { renderComponentWithHook } from "@evg-ui/lib/test_utils/hooks";
 import { ToastProvider, useToastContext } from ".";
 import { RenderFakeToastContext } from "./__mocks__";
 import { TOAST_TIMEOUT } from "./constants";
